requirements.txt
accelerate==1.4.0
acres==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
asgiref==3.8.1
astor==0.8.1
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.1
captcha==0.6.0
celery==5.4.0
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==2.1.0
cloudpickle==3.1.1
coloredlogs==15.0.1
cryptography==42.0.5
datasets==3.2.0
evaluate==0.4.3
frozenlist==1.5.0
google-auth==2.29.0
grpcio==1.67.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.1
httpx
Jinja2==3.1.5
Jinja2==3.1.5
langchain==0.3.17
langchain-community==0.3.16
langchain==0.3.17
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
opencv-python==4.9.0.80
openai==1.23.2
opentelemetry-api==1.30.0
orjson==3.10.15
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydot==3.0.2
pypdf==5.2.0
python-multipart==0.0.9
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.4
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
xformers

